Se agrego: -Busquedas con arboles -Registro -Registrar -Editar valores

+++ Proyecto_Estructuras/Models/paciente.cs
@@ -9,7 +9,7 @@ namespace Proyecto_Estructuras.Models
     {
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int DPI_CUI { get; set; }
+        public long DPI_CUI { get; set; }
         public string Departamento { get; set; }
         public string Municipio_residencia { get; set; }
         public int Edad { get; set; }
